Setup three states for the SetupScreen, the intial state where user's enter there names, the dice roll state, where user's roll their dice, and the team select screen, where the winner of the dice roll chooses their team. All states inlcude buttons that trigger the next state to occur.
Two main functions outside of the event functions were created as well. The first being createPlayerHands which sets the players hands based off what team the winner chose. The second was setPlayers which takes in both player objects from the MainScreen Form and assigns them to the player objects on the SetupScreen.

Finish this off by including a Start Game button which creates the player's hands, and passes those players to back to the MainForm so that they can be passed to the GameScreen UserControl.