Created CardSlot UserControl for event driven controls so that I can register if a user clicks on a CardSlot in their hand. This includes three attributes: Card _card, bool _isSelected, and event Action<CardSlot> CardClicked.
The constructor for CardSlot adds the controls to the CardSlot_Click event handler so that when they are clicked they trigger that event.

the LoadCard method takes in a Card object and sets the card object passed to the CardSlot object's Card Object. This allows us to pass the selected card from the players hand to the CardSlot which then updates the CardSlot to show the Card from the players hand.

The ClearSlot method sets the local card object to null and clears out the data in the CardSlot object.

The SetSelected method sets the field _isSelected to true and then changes the BackColor of the CardSlot to gold.

The ClearSelected method does the opposite of the SetSelected method with the BackColor set to transparent.