Created BoardSlot UserControl for event logic uses.
Includes three fields, Card _card, bool _isEmpty, event Action<BoardSlot> SlotClicked

The constructor adds the UI elements Clicks to the BoardSlot_Click function so that the function is called when each UI item is clicked.

There is a public property, IsEmpty that returns isEmpty

There are three total methods.
The first LoadCardInSlot, takes a Card object, sets that card object to the BoardSlots local Card object and passes the Cards data from one object to the other. isEmpty is also set to false when this is called.

The second method, ClearSlot, sets the Card object to null, isEmpty to true, and clears the data out of the UI elements.

The final method is an event method that invokes when the Slot is clicked. And sends the event to any objects added to the SlotClicked event.