Added eight fields, one for the Game object, two for each players Hand Slots, two for each players Active board slots, remainingTime for the timer, _selectedCard, and _selectedCardSlot
There are seven methods that were added to this class. OnCardSelected, takes in a CardSlot and checks which player's turn it is and as long as the players hand contains that card, set the selectedCardSlot to the slot being passed to the function and set the selectedCard to the card in said slot.

The next method is OnBoardSlotClicked which takes in a BoardSlot object and based on if the player's is selecting their respecitve active slot, consume the mana that the card shows, load the selected card in the slot, remove the card from the player's hand, reload UI, then set the selected items to null.

The next method is LoadGame, this takes in a Game object and sets the GameScreens local game object to the passed in object. Then remainingTime is set back to 30, the timer starts, and the label to whose turn it is is set to Player 1's name.

The method LoadUI passes the Game objects data and player data to the UI elements on the GameScreen.

The method buttonP1EndTurn_Click is an event method that listens for a click of player one's end turn button and once it is hit, run the endTurn game method, disable player one's end turn button and enable player two's. then we load the UI, update player 2 name to be shown as the active turn holder, reset remaining time, and start the timer.

The method buttonP2EndTurn is the same as the P1 end turn method except in reverse.

Finally, the timerTurn_Tick event method, decrements the remaining time by one each tick, and if the time hits 0, it ends the active players turn.